pyproject.toml
[build-system]
requires = ["setuptools>=68"]         # PEP 660 editable support ke liye naya setuptools
build-backend = "setuptools.build_meta"


[project]
name = "qna-assistant-using-dspy-and-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "dspy-ai>=2.4.0",
    "ipykernel>=6.30.1",
    "langchain-core>=0.2.0",
    "langgraph>=0.1.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pydantic>=2.0.0",
    "pyppeteer>=2.0.0",
    "rank-bm25>=0.2.2",
    "rich>=13.7.0",
    "scikit-learn>=1.3.0",
]

authors = [{ name = "Praveena Raju" }]


[tool.setuptools.packages.find]
where = ["agent"]
include = ['*']
